Added sign support, moved some classes and methods
Added Sign Support! Vanilla does not have a ScreenHandler for signs as its normally client sided only, because of this I have added a sign screen handler so we can keep track of it. It also means we have to send a ghost sign block + entity to the player which contains the properties. This block will be removed when the sign is closed. I also put the onTick method to GuiInterface so its universal to all guis.